requirements.txt
python-telegram-bot==12.0.0.b1
requests==2.21.0
python-dotenv==0.10.1
uptime==3.0.1
pytz==2019.1